move unit tests into their respective folders
remove set command and replace with configure to set the
configuration on an app/route
