--- commands/set.go
@@ -1,17 +1,0 @@
-package commands
-
-import (
-	"github.com/urfave/cli"
-)
-
-func SetCommand() cli.Command {
-	return cli.Command{
-		Name:        "set",
-		Aliases:     []string{"s"},
-		Usage:       "set command",
-		Category:    "MANAGEMENT COMMANDS",
-		Hidden:      false,
-		ArgsUsage:   "<command>",
-		Subcommands: GetCommands(SetCmds),
-	}
-}
